@@ -6,12 +6,12 @@ namespace Façade
 {
     public class Generator
     {
-        private static readonly Random random = new Random();
+        private static readonly Random Random = new();
 
-        public List<int> Generate(int count)
+        public static List<int> Generate(int count)
         {
             return Enumerable.Range(0, count)
-                .Select(_ => random.Next(1, 6))
+                .Select(_ => Random.Next(1, 6))
                 .ToList();
         }
     }
