@@ -1,27 +1,27 @@
-namespace BuilderFacets
+namespace Builder
 {
     public class PersonJobBuilder : PersonBuilder
     {
         public PersonJobBuilder(Person person)
         {
-            this.person = person;
+            Person = person;
         }
 
         public PersonJobBuilder At(string companyName)
         {
-            person.CompanyName = companyName;
+            Person.CompanyName = companyName;
             return this;
         }
 
         public PersonJobBuilder AsA(string position)
         {
-            person.Position = position;
+            Person.Position = position;
             return this;
         }
 
         public PersonJobBuilder Earning(int annualIncome)
         {
-            person.AnnualIncome = annualIncome;
+            Person.AnnualIncome = annualIncome;
             return this;
         }
     }
