@@ -2,10 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using Façade;
 using NUnit.Framework;
 
-namespace Coding.Exercise.Tests
+namespace Façade
 {
     [TestFixture]
     public class TestSuite
@@ -24,15 +23,13 @@ namespace Coding.Exercise.Tests
         public void TestSizeThree()
         {
             var gen = new MagicSquareGenerator();
-            var square = gen.Generate(3);
+            var square = MagicSquareGenerator.Generate(3);
 
             Console.WriteLine(SquareToString(square));
 
             var v = new MyVerifier(); // prevents cheating :)
-            Assert.IsTrue(v.Verify(square),
+            Assert.IsTrue(MyVerifier.Verify(square),
                 "Verification failed: this is not a magic square");
         }
     }
-}
-
 }
