@@ -4,7 +4,7 @@ namespace Factories
     {
         public static Point NewCartesianPoint(float x, float y)
         {
-            return new Point(x, y); // needs to be public
+            return new(x, y); // needs to be public
         }
     }
 }
