@@ -2,7 +2,7 @@
 
 namespace Flyweight
 {
-    internal class Program
+    public static class Program
     {
         private static void Main(string[] args)
         {
