@@ -1,15 +1,13 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 
 namespace Singleton
 {
     public class SingletonRecordFinder
     {
-        public int TotalPopulation(IEnumerable<string> names)
+        public static int TotalPopulation(IEnumerable<string> names)
         {
-            var result = 0;
-            foreach (var name in names)
-                result += SingletonDatabase.Instance.GetPopulation(name);
-            return result;
+            return names.Sum(name => SingletonDatabase.Instance.GetPopulation(name));
         }
     }
 }
